edits

+++ include/universal/blas/solvers/backsub.hpp
@@ -17,9 +17,9 @@ Vector backsub(const Matrix& A, const Vector& b) {
 	size_t n = size(b);
     Vector x(n);
     
-	for (int i = n-1; i >=0 ;--i){
+	for (size_t i = n-1; i >=0 ;--i){
         Scalar y = 0.0;
-        for (int j = i; j < n; ++j){
+        for (size_t j = i; j < n; ++j){
             y += A(i,j)*x(j);
         }
         x(i) = (b(i) - y)/A(i,i);
@@ -37,9 +37,9 @@ vector<posit<nbits,es>> backsub(const matrix<posit<nbits,es>> & A, const vector<
     constexpr unsigned capacity = 10;
 
     Vector x(n);
-	for (int i = n-1; i >=0; --i){
+	for (size_t i = n-1; i >=0; --i){
         quire<nbits,es,capacity> q{0};
-        for (int j = i; j < n; ++j){
+        for (size_t j = i; j < n; ++j){
             q += quire_mul(A(i,j), x(j));
         }
         posit<nbits,es> y;
